+++ runtime/mandala/src/lib.rs
@@ -1172,7 +1172,7 @@ impl module_nft::Trait for Runtime {
 }
 
 impl orml_nft::Trait for Runtime {
-	type ClassId = u64;
+	type ClassId = u32;
 	type TokenId = u64;
 	type ClassData = module_nft::ClassData;
 	type TokenData = module_nft::TokenData;
